pyproject.toml
[project]
name = "tushare-stock-mcp"
version = "0.1.0"
description = "Tushare 股票分析 MCP 服务"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "tushare>=1.2.89",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
managed = true